Changes name of the original library.rb and library_spec.rb tto book.rb and book_spec.rb, creates new library lib- and spec-file for the new Library class.
